Regular display fixes including refresh tweaks

+++ shared-module/displayio/TileGrid.h
@@ -43,7 +43,7 @@ typedef struct {
     uint16_t pixel_width;
     uint16_t pixel_height;
     uint16_t bitmap_width_in_tiles;;
-    uint8_t tiles_in_bitmap;
+    uint16_t tiles_in_bitmap;
     uint16_t width_in_tiles;
     uint16_t height_in_tiles;
     uint16_t tile_width;
